requirement.txt
Flask==2.2.3
Pillow==9.4.0
numpy==1.24.2
pandas==1.5.3

